Fix: MOTW brakes sim_coffee.html
The "mark of the web" (MOTW) was previously added to the sim*.html files
to allow automatic execution on IE (without the "allow blocked content"
button prompt).  This broke the sim_coffee.html file on IE.  Apparently,
coffee-script.js needs to be explicitly unblocked to allow loading of
.coffee files.  This was originally done by having the MOTW in the
template_body.html file so that it is automatically included in all
sim*.html files.

The "mark of the web" was moved from the template_body.html to both the
template_head.html and js_head.html (but not coffee_head.html).  This
way, the MOTW is explicitly included in sim.html and sim_js.html, but
not sim_coffee.html.